@@ -1,5 +1,5 @@
 {
-  "name": "react-xterm",
+  "name": "@termftp/react-xterm",
   "version": "3.6.0",
   "description": "react wrapper for xterm.js",
   "main": "./dist/react-xterm.js",
@@ -11,30 +11,31 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/farfromrefug/react-xterm"
+    "url": "https://github.com/TermFTP/react-xterm"
   },
-  "author": "Martin Guillon",
+  "author": "TermFTP",
   "license": "ISC",
   "peerDependencies": {
     "react": "^16.3.1"
   },
   "dependencies": {
-    "classnames": "^2.2.6",
-    "xterm": "3.6.0",
-    "chai": "^4.2.0",
+    "chai": "^4.3.4",
+    "classnames": "^2.3.1",
     "express-ws": "^4.0.0",
-    "node-pty": "^0.9.0"
+    "node-pty": "^0.9.0",
+    "xterm": "3.6.0"
   },
   "devDependencies": {
-    "@types/node": "^7.0.60",
-    "@types/react": "^16.3.10",
+    "@types/classnames": "^2.3.1",
+    "@types/node": "^7.10.14",
+    "@types/react": "^16.14.5",
     "@types/xterm": "^3.0.0",
     "lodash.throttle": "^4.1.1",
-    "react": "^16.3.1",
-    "react-dom": "^16.3.1",
-    "react-resizable": "^1.7.5",
-    "ts-loader": "^4.2.0",
-    "typescript": "^2.8.1",
-    "webpack": "^4.5.0"
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
+    "react-resizable": "^1.11.1",
+    "ts-loader": "^4.5.0",
+    "typescript": "^2.9.2",
+    "webpack": "^4.46.0"
   }
 }
